fix concurrent bug and should not insert conflict command into spec pool
The bug will happen in the following situation:
1. A client proposes a command to all servers.
2. The leader receives the command, sync to others, and committed the command.
3. A follower it notices that the command is committed and executes it. When the execution is over, it will need to remove the command from spec pool. However, since the follower may have not yet received the proposal from the client, it will find no such command in the spec pool and panic.
Suggested fix:
When a follower tries to remove a non-exist command from spec pool, it will register the command id in an internal buffer(ready). When the proposal is finally received, it will directly remove the command. The ready pool needs gc periodically since the proposal may never arrive.