fix bug discovered in tests
fix leader should set leader id to itself when it becomes leader
fix slow path might be blocked by adding a timeout
fix client might be blocked: If a server loses contact with its leader, it will update its term for election. Since other servers are all right, the election will not succeed. But if the client learns about the new term and updates its term to it, it will never get the true leader. Now we will reset client's term only when the server has a leader id.
fix that since the follower timeout is changed, we need to wait for a longer time for leader to be elected